Change SessionId type from long to Int128 in SessionData model and update RedisSessionService to parse sessionId accordingly

+++ Models/Models.cs
@@ -4,7 +4,7 @@ namespace requete.Models;
 
 public class SessionData
 {
-    public long SessionId { get; set; }
+    public Int128 SessionId { get; set; }
     public long UserId { get; set; }
     public string? UserFullname { get; set; }
 }
